+++ chainercv/utils/__init__.py
@@ -1,4 +1,5 @@
 from chainercv.utils.bbox.bbox_overlap import bbox_overlap  # NOQA
+from chainercv.utils.bbox.non_maximum_suppression import non_maximum_suppression  # NOQA
 from chainercv.utils.download import cached_download  # NOQA
 from chainercv.utils.download import extractall  # NOQA
 from chainercv.utils.extension import check_type  # NOQA
